fastapi-backend/requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
av==13.1.0
bcrypt==4.1.3
blis==1.0.1
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpeg==1.4
filelock==3.16.1
fsspec==2024.10.0
geojson==2.5.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.7
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
limits==3.12.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.4.2
numpy==2.0.2
openai==1.30.5
orjson==3.10.3
packaging==24.0
passlib==1.7.4
pillow==10.4.0
preshed==3.0.9
psycopg2==2.9.9
pyasn1==0.6.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyneuphonic==1.1.3
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.1
regex==2024.11.6
requests==2.32.3
rich==13.7.1
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.0
shellingham==1.5.4
six==1.16.0
slowapi==0.1.9
smart-open==7.0.5
sniffio==1.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
starlette==0.37.2
sympy==1.13.1
thinc==8.3.2
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
tqdm==4.66.4
transformers==4.46.2
typer==0.12.3
typing_extensions==4.12.0
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
wasabi==1.1.3
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
